# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Case {
  id: number;
  title: string;
  description: string;
  /** 1=easy, 2=medium, 3=hard, 4=expert, 5=master */
  difficulty: number;
  isPublished: boolean;
  isPremium: boolean;
  /** @nullable */
  coverImage?: string | null;
  location: string;
  crimeType: string;
  /** Points awarded for solving */
  reward: number;
  createdAt: string;
}

export interface Suspect {
  id: number;
  caseId: number;
  name: string;
  role: string;
  backstory: string;
  /** @nullable */
  photoUrl?: string | null;
  isGuilty: boolean;
  /** 1-10, how evasive the suspect is */
  deceptionLevel: number;
  personality: string;
}

export interface CaseDetail {
  id: number;
  title: string;
  description: string;
  difficulty: number;
  isPublished: boolean;
  isPremium: boolean;
  /** @nullable */
  coverImage?: string | null;
  location: string;
  crimeType: string;
  reward: number;
  createdAt: string;
  suspects: Suspect[];
  evidenceList: string[];
}

export interface CaseInput {
  title: string;
  description: string;
  difficulty: number;
  isPremium: boolean;
  location: string;
  crimeType: string;
  reward: number;
  evidenceList?: string[];
  adminKey: string;
}

export interface CaseUpdate {
  title?: string;
  description?: string;
  difficulty?: number;
  isPremium?: boolean;
  isPublished?: boolean;
  location?: string;
  crimeType?: string;
  reward?: number;
  evidenceList?: string[];
  adminKey?: string;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface InterrogationInput {
  suspectId: number;
  caseId: number;
  sessionId: string;
  message: string;
  history?: ChatMessage[];
}

export interface InterrogationResponse {
  response: string;
  /** calm, nervous, angry, defensive, crying */
  suspectMood: string;
}

export interface GuessInput {
  caseId: number;
  sessionId: string;
  suspectId: number;
}

export interface GuessResult {
  correct: boolean;
  message: string;
  pointsEarned: number;
}

export interface Player {
  id: number;
  sessionId: string;
  name: string;
  hasPaid: boolean;
  points: number;
  paymentExempt: boolean;
  createdAt: string;
}

export interface PlayerInput {
  sessionId: string;
  name: string;
}

export interface PlayerUpdate {
  hasPaid?: boolean;
  paymentExempt?: boolean;
  adminKey?: string;
}

export interface PlayerProgress {
  id: number;
  playerId: number;
  caseId: number;
  isSolved: boolean;
  attempts: number;
  /** @nullable */
  solvedAt?: string | null;
}

export interface PurchaseInput {
  sessionId: string;
  caseId: number;
}

export interface PurchaseResult {
  success: boolean;
  message: string;
}

export interface GameSettings {
  id: number;
  paymentEnabled: boolean;
  /** Amount in cents */
  paymentAmount: number;
  /** Number of free cases before payment required */
  freeTrialCases: number;
}

export interface GameSettingsUpdate {
  paymentEnabled?: boolean;
  paymentAmount?: number;
  freeTrialCases?: number;
  adminKey?: string;
}

export interface AdminStats {
  totalPlayers: number;
  totalCases: number;
  publishedCases: number;
  totalSolvedCases: number;
  paidPlayers: number;
}

export type GetPlayerParams = {
sessionId: string;
};

export type GetPlayerProgressParams = {
sessionId: string;
};

export type AdminListCasesParams = {
adminKey: string;
};

export type AdminListPlayersParams = {
adminKey: string;
};

export type AdminGetStatsParams = {
adminKey: string;
};