# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalPlayers: number;
  totalCases: number;
  publishedCases: number;
  totalSolvedCases: number;
  paidPlayers: number;
}