# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  title: string;
  description: string;
  /** 1=easy, 2=medium, 3=hard, 4=expert, 5=master */
  difficulty: number;
  isPublished: boolean;
  isPremium: boolean;
  /** @nullable */
  coverImage?: string | null;
  location: string;
  crimeType: string;
  /** Points awarded for solving */
  reward: number;
  createdAt: string;
}