# lib/api-zod/src/generated/types/caseDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Suspect } from './suspect';

export interface CaseDetail {
  id: number;
  title: string;
  description: string;
  difficulty: number;
  isPublished: boolean;
  isPremium: boolean;
  /** @nullable */
  coverImage?: string | null;
  location: string;
  crimeType: string;
  reward: number;
  createdAt: string;
  suspects: Suspect[];
  evidenceList: string[];
}