# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  title: string;
  description: string;
  difficulty: number;
  isPremium: boolean;
  location: string;
  crimeType: string;
  reward: number;
  evidenceList?: string[];
  adminKey: string;
}