# lib/api-zod/src/generated/types/gameSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSettings {
  id: number;
  paymentEnabled: boolean;
  /** Amount in cents */
  paymentAmount: number;
  /** Number of free cases before payment required */
  freeTrialCases: number;
}