# lib/api-zod/src/generated/types/gameSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSettingsUpdate {
  paymentEnabled?: boolean;
  paymentAmount?: number;
  freeTrialCases?: number;
  adminKey?: string;
}