# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  sessionId: string;
  name: string;
  hasPaid: boolean;
  points: number;
  paymentExempt: boolean;
  createdAt: string;
}