# lib/api-zod/src/generated/types/playerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerUpdate {
  hasPaid?: boolean;
  paymentExempt?: boolean;
  adminKey?: string;
}