# lib/api-zod/src/generated/types/suspect.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Detective Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Suspect {
  id: number;
  caseId: number;
  name: string;
  role: string;
  backstory: string;
  /** @nullable */
  photoUrl?: string | null;
  isGuilty: boolean;
  /** 1-10, how evasive the suspect is */
  deceptionLevel: number;
  personality: string;
}